% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{batch_check}
\alias{batch_check}
\title{Check Batch Setup}
\usage{
batch_check(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
An invisible copy of the values of .argsbatchr.rds.
}
\description{
Check Batch Setup
}
