% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{batch_cleanup}
\alias{batch_cleanup}
\title{Cleanup Batch Processing}
\usage{
batch_cleanup(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
An invisible character vector of the files that have been processed.
}
\description{
Cleanup Batch Processing
}
\seealso{
\code{\link{batch_process}()}
}
