% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completed.R
\name{batch_completed}
\alias{batch_completed}
\title{Batch Completed?}
\usage{
batch_completed(path = ".", failed = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{failed}{A logical scalar specifying whether to exclude (FALSE),
include (NA), or only include (TRUE) files that have thus far failed to process.}
}
\value{
A flag specifying whether batch processing is complete.
}
\description{
Tests if there are any remaining files to batch process.
}
\details{
By default, files that were not successfully processed are 
excluded.
}
\seealso{
\code{\link{batch_process}()}
}
