% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{batch_config}
\alias{batch_config}
\title{Configure Batch Processing}
\usage{
batch_config(FUN, path = ".", pattern = ".*", recursive = FALSE, ...)
}
\arguments{
\item{FUN}{A function to process the files. 
The function's first argument must be a string of the path 
to a single file.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
An invisible character vector of the paths to the files to
be processed.
}
\description{
Sets up files for batch processing by specifying the
files to process and the function to use and saving the information in
a .batchr.rds configuration file.
}
\details{
Set up requires that path does not already include a .batchr.rds 
configuration file (and if recursive = FALSE nor do any of its subdirectories).

Hidden files are not included in those for batch processing to
prevent accidental modification of system files.
}
\seealso{
\code{\link{batch_read_config}()}
}
