% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{batch_config_files}
\alias{batch_config_files}
\title{Configuration Files}
\usage{
batch_config_files(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{recursive}{A flag specifying whether to recurse into subdirectories
when cleaning up.}
}
\value{
A character vector of the names of the configuration file(s).
}
\description{
Configuration Files
}
\seealso{
\code{\link{batch_process}()}
}
