% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_config_read}
\alias{batch_config_read}
\title{Read Configuration File}
\usage{
batch_config_read(path)
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}
}
\value{
A named list of the configuration values.
}
\description{
Reads the values in the 
configuration file created by \code{\link{batch_config}()}.
}
\seealso{
\code{\link{batch_process}()} and \code{\link{batch_read_log}()}
}
