% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{batch_config_update}
\alias{batch_config_update}
\title{Updates Batch Processing Configuration}
\usage{
batch_config_update(fun, path = ".", ...)
}
\arguments{
\item{fun}{A function to process each of the files. 
\code{fun}'s first argument should be a string of the path to a single file.
If processing is unsuccessful \code{fun} should return FALSE 
or throw an error (error messages are caught and automatically logged).}

\item{path}{A string of the path to the directory with the files for processing.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
An invisible character vector of the paths to the files 
remaining to be processed.
}
\description{
Updates the function that was provided when a directory was configured 
using \code{\link{batch_config}()}.
}
\details{
\code{batch_config_update()} is useful if a new version of the files is required
to successfully process some of the files.

It should be noted that \code{batch_config_update()} does not alter the 
configuration time.

In order to process previously failed files \code{\link{batch_run}()}
should be called with \code{failed = NA} or \code{failed = TRUE}.
}
\seealso{
\code{\link{batch_process}()} and \code{\link{batch_config}()}
}
