% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{batch_files}
\alias{batch_files}
\title{Batch File Names}
\usage{
batch_files(path = ".", processed = NA)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{processed}{A logical scalar specifying whether to return the names of 
all the file (default), the number processed (TRUE) or the number remaining (FALSE).}
}
\value{
A character vector of the paths to the files.
}
\description{
Batch File Names
}
