% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub.R
\name{batch_gsub}
\alias{batch_gsub}
\title{Batch File Text Replacement}
\usage{
batch_gsub(pattern, replacement, path = ".", regexp = "[.](R|r)$",
  recurse = FALSE, parallel = FALSE, ask = getOption("batchr.ask",
  TRUE))
}
\arguments{
\item{pattern}{A string of the regular expression to match.}

\item{replacement}{A string of the replacement text.}

\item{path}{A string of the path to the directory with the files for processing.}

\item{regexp}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recurse}{A flag specifying whether to recurse into path's subdirectories.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}

\item{ask}{A flag specifying whether to ask before starting to process the files.}
}
\value{
An invisible flag indicating whether all the files were 
successfully processed.
}
\description{
Uses \code{\link{batch_process}()} and \code{\link{gsub_file}()} to
perform batch text file replacement.
By default it replaces text in all .R and .r files in the working directory.
}
\seealso{
\code{\link{batch_process}()}
}
