% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{batch_log_files}
\alias{batch_log_files}
\title{Log Files}
\usage{
batch_log_files(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}
}
\value{
A character vector of the names of the log file(s).
}
\description{
Log Files
}
\seealso{
\code{\link{batch_process}()}
}
