% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch_process}
\alias{batch_process}
\title{Batch Process}
\usage{
batch_process(FUN, path = ".", pattern = ".*", recursive = FALSE,
  ..., parallel = TRUE)
}
\arguments{
\item{FUN}{A function to process the files. The first argument should be a string of the absolute path to a single file.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of the regular expression to match to file names.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{...}{Additional arguments passed to FUN.}

\item{parallel}{A flag specifying whether to process the files in parallel.}
}
\value{
An invisible character vector of the processed files.
}
\description{
Performs set up, processing and clean up.
It can be halted during processing and restarted using \code{\link{batch_restart}()}
}
