% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_read_args}
\alias{batch_read_args}
\title{Batch Read Arguments}
\usage{
batch_read_args(path)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A named list of the argument values.
}
\description{
Reads the arguments to batchr that were stored in the .argsbatchr.rds file
when the project was setup.
}
