% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_read_setup}
\alias{batch_read_setup}
\title{Batch Read Setup}
\usage{
batch_read_setup(path)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A named list of the argument values.
}
\description{
Reads the values in the .batchr_setup.rds file which was created 
when the project was setup.
}
