% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{batch_report}
\alias{batch_report}
\title{Batch Report}
\usage{
batch_report(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}
}
\description{
Outputs a report of the status of individual files to the console.
}
\seealso{
\code{\link{batch_file_status}()}
}
