% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{batch_reset}
\alias{batch_reset}
\title{Batch Reset}
\usage{
batch_reset(path, recursive = NA)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{recursive}{A logical scalar specifying whether to leave the recursive
setting unchanged (default) or reset it to be TRUE or FALSE.}
}
\value{
An invisible character vector of the paths to the files to
be processed.
}
\description{
Batch Reset
}
