% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seeds.R
\name{batch_seeds}
\alias{batch_seeds}
\title{L'Ecuyer-CMRG Seeds}
\usage{
batch_seeds(files = batch_files_remaining())
}
\arguments{
\item{files}{A character vector of the names of the files.}
}
\value{
A named list of the L'Ecuyer-CMRG seed for each file name.
}
\description{
Generates a named list of L'Ecuyer-CMRG seeds.
}
\examples{
batch_seeds(c("a", "b"))
}
