% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{batch_setup}
\alias{batch_setup}
\title{Setup Batch Processing of Files}
\usage{
batch_setup(FUN, path = ".", pattern = ".*", recursive = FALSE, ...)
}
\arguments{
\item{FUN}{A function to process the files. The first argument should be a string of the absolute path to a single file.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of the regular expression to match to file names.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
An invisible character vector of the paths to the files to
be processed.
}
\description{
Setup Batch Processing of Files
}
