% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{batch_setup_files}
\alias{batch_setup_files}
\title{List Set Up File Names}
\usage{
batch_setup_files(path = ".", recursive = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}
}
\value{
A character vector of the file names.
}
\description{
List Set Up File Names
}
