% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{batch_summary}
\alias{batch_summary}
\title{Batch Summary}
\usage{
batch_summary(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}
}
\description{
Writes a summary of individual file status etc to the console.
}
\seealso{
\code{\link{batch_file_status}()}
}
