% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcount.R
\name{rcount}
\alias{rcount}
\title{Random Count}
\usage{
rcount(n = 1L)
}
\arguments{
\item{n}{A count of the number of random counts to generate.}
}
\value{
An integer vector of random counts.
}
\description{
Generates a platform independent random count between 0 and 
2,147,483,647.
It is useful for specifying values for \code{\link{set.seed}()}.
}
\examples{
set.seed(45)
rcount()
set.seed(45)
rcount(n = 2L)
}
