% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rinteger}
\alias{rinteger}
\title{Random Integer}
\usage{
rinteger(n = 1)
}
\arguments{
\item{n}{A non-negative whole number of the number of random integers to generate.}
}
\value{
A integer vector.
}
\description{
Generates random integers between -2147483647L and 2147483647L.
}
\examples{
rinteger()
}
