% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{batch_cleanup}
\alias{batch_cleanup}
\title{Cleanup Batch Processing}
\usage{
batch_cleanup(path = ".", force = FALSE, remaining = FALSE,
  failed = NA, silent = FALSE, recursive = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}

\item{force}{A flag specifying whether to remove configuration and 
log files even if there are files remaining to be processed.}

\item{remaining}{When \code{force = TRUE} a flag specifying 
whether to remove files that have not yet been processed.}

\item{failed}{When \code{remaining = TRUE} a logical scalar specifying 
whether to remove all remaining files (NA), 
only those files that failed to process (TRUE),
or only untested files (FALSE).}

\item{silent}{A flag specifying whether to suppress warnings about
directories that could not be cleaned.}

\item{recursive}{A flag specifying whether to recurse into subdirectories
when cleaning up.}
}
\value{
A named logical vector indicating which directories 
were successfully cleaned up.
}
\description{
Removes configuration file created by \code{\link{batch_config}()} 
and log file created by \code{\link{batch_run}()}.
}
\seealso{
\code{\link{batch_process}()}
}
