% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completed.R
\name{batch_completed}
\alias{batch_completed}
\title{Batch Completed?}
\usage{
batch_completed(path = ".", failed = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}

\item{failed}{A logical scalar specifying how to treat files 
that previously failed to process. If FALSE (the default) failed files 
are excluded, if NA they are included and if TRUE they are only included.}
}
\value{
A flag specifying whether batch processing is complete.
}
\description{
Tests if there are any remaining files to process as listed
by \code{\link{batch_files_remaining}()}.
}
\details{
By default, files that previously failed to process are excluded.
}
\seealso{
\code{\link{batch_process}()}
}
