% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{batch_config}
\alias{batch_config}
\title{Configure Batch Processing}
\usage{
batch_config(FUN, path = ".", pattern = ".*", recursive = FALSE, ...)
}
\arguments{
\item{FUN}{A function to process each of the files.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{...}{Additional arguments passed to FUN.}
}
\value{
An invisible character vector of the paths to the files to
be processed.
}
\description{
Configures a directory for batch file processing.
}
\seealso{
\code{\link{batch_process}()}
}
