% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{batch_config}
\alias{batch_config}
\title{Configure Batch Processing}
\usage{
batch_config(fun, path = ".", regexp = ".*", recurse = FALSE, ...)
}
\arguments{
\item{fun}{A function to process each of the files. 
\code{fun}'s first argument should be a string of the path to a single file.
If processing is unsuccessful \code{fun} should return FALSE 
or throw an error (error messages are automatically logged).}

\item{path}{A string of the path to the directory with the files for processing.}

\item{regexp}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recurse}{A flag specifying whether to recurse into path's subdirectories.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
An invisible character vector of the paths to the files to
that will be processed when \code{\link{batch_run}()} is called.
}
\description{
Configures a directory for batch file processing by \code{\link{batch_run}()}.
}
\details{
\code{batch_config()} creates a hidden configuration file in path 
named '.batchr.rds'.

The contents of the file can be read using
\code{\link{batch_config_read}()} 
or updated using \code{\link{batch_config_update}()}.
}
\seealso{
\code{\link{batch_process}()} and \code{\link{batch_run}()}
}
