% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub.R
\name{batch_gsub}
\alias{batch_gsub}
\title{Batch File Text Replacement}
\usage{
batch_gsub(pattern_gsub, replacement_gsub, path = ".",
  pattern = "[.](R|r)$", recursive = FALSE, parallel = FALSE)
}
\arguments{
\item{pattern_gsub}{A string of the regular expression to match.}

\item{replacement_gsub}{A string of the replacement text.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}
}
\value{
An invisible flag indicating whether all the files were 
successfully processed.
}
\description{
Uses \code{\link{batch_process}()} and \code{\link{gsub_file}()} to
replace text in by default .R and .r files in the working directory.
}
\seealso{
\code{\link{batch_process}()}
}
