% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch_process}
\alias{batch_process}
\title{Batch File Processing}
\usage{
batch_process(fun, path = ".", regexp = ".*", recurse = FALSE,
  parallel = FALSE, progress = TRUE, force = TRUE,
  ask = getOption("batchr.ask", TRUE), ...)
}
\arguments{
\item{fun}{A function to process each of the files. 
\code{fun}'s first argument should be a string of the path to a single file.
If processing is unsuccessful \code{fun} should return FALSE 
or throw an error (error messages are caught and automatically logged).}

\item{path}{A string of the path to the directory with the files for processing.}

\item{regexp}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recurse}{A flag specifying whether to recurse into path's subdirectories.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}

\item{progress}{A string specifying the type of output to log to the console
when \code{parallel = FALSE}. 
The possible values are "all", "failures" or "none".}

\item{force}{A flag specifying whether to delete configuration and 
log files even if there are files remaining to be processed.}

\item{ask}{A flag specifying whether to ask before starting to process the files.}

\item{...}{Additional arguments passed to \code{fun}.}
}
\value{
An invisible flag indicating whether all the files where 
successfully processed.
}
\description{
Performs batch processing of files in a directory using the 
\code{\link{batch_config}()}, \code{\link{batch_run}()}
and \code{\link{batch_cleanup}()} functions.
For more control the user should call these three functions instead.
}
\details{
\code{\link{batch_gsub}()} provides a simple wrapper for
batch text replacement.
}
\seealso{
\code{\link{batch_config}()}, \code{\link{batch_run}()}
and \code{\link{batch_cleanup}()}
}
