% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{batch_process}
\alias{batch_process}
\title{Batch Process}
\usage{
batch_process(FUN, path = ".", pattern = ".*", recursive = FALSE,
  ..., parallel = TRUE)
}
\arguments{
\item{FUN}{A function to process the files. 
The function's first argument must be a string of the path 
to a single file.}

\item{path}{A string of the path to the directory with the files.}

\item{pattern}{A string of a regular expression. Only non-hidden file names 
which match the regular expression will be batch processed.}

\item{recursive}{A flag specifying whether to recurse into sub directories.}

\item{...}{Additional arguments passed to FUN.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}
}
\value{
A flag specifying whether it was able to process.
}
\description{
Performs set up, processing and clean up.
}
