% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_read_config}
\alias{batch_read_config}
\title{Read Configuration File}
\usage{
batch_read_config(path)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A named list of the argument values.
}
\description{
Reads the values in the .batchr.rds configuration file which was created 
when the project was set up.
}
\seealso{
\code{\link{batch_process}()}
}
