% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_read_log}
\alias{batch_read_log}
\title{Read Log File}
\usage{
batch_read_log(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A tibble with four columns.
type is a character vector indicating SUCCESS or FAILURE,
time is a POSIXct of the datetime of processing in UTC, 
file is a character vector of the file name and 
error is a character vector of the error message (or NA if no error).
}
\description{
Reads the values in the .batchr.log processing failure file.
}
\seealso{
\code{\link{batch_process}()}
}
