% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{batch_read_log}
\alias{batch_read_log}
\title{Read Log File}
\usage{
batch_read_log(path = ".")
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A tibble with an ordered factor of the \code{level} (WARN, ERROR, FATAL), 
a POSIXct of the time of failure in UTC and a character vector of the file name.
}
\description{
Reads the values in the .batchr.log processing failure file.
}
\seealso{
\code{\link{batch_process}()}
}
