% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-setup.R
\name{batch_read_setup}
\alias{batch_read_setup}
\title{Batch Read Setup}
\usage{
batch_read_setup(path)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}
}
\value{
A named list of the argument values.
}
\description{
Reads the arguments to batchr that were stored in the 
.batchr_setup.rds file when the project was setup.
}
