% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{batch_reconfig}
\alias{batch_reconfig}
\title{Reconfigures Batch Processing}
\usage{
batch_reconfig(fun, path = ".", ...)
}
\arguments{
\item{fun}{A function to process each of the files.}

\item{path}{A string of the path to the directory with the files.}

\item{...}{Additional arguments passed to fun.}
}
\value{
An invisible character vector of the paths to the files 
(failed and untested) remaining to be processed.
}
\description{
Updates the batch processing function in the configuration file.
}
\seealso{
\code{\link{batch_process}()}
}
