% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{batch_run}
\alias{batch_run}
\title{Runs Batch Processing}
\usage{
batch_run(path = ".", failed = FALSE, parallel = FALSE,
  progress = !parallel, files = NULL, seed = NULL,
  ask = getOption("batchr.ask", TRUE))
}
\arguments{
\item{path}{A string of the path to the directory with the files for processing.}

\item{failed}{A logical scalar specifying how to treat files 
that previously failed to process. If FALSE (the default) failed files 
are excluded, if NA they are included and if TRUE they are only included.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}

\item{progress}{A logic scalar specifying whether to 
write logging information to the console (TRUE or FALSE) or only to write
logging information to the console for failed attempts (NA).}

\item{files}{A character vector of the remaining files to process.
If \code{NULL} then \code{files} is as \code{batch_files_remaining(path, failed)}.}

\item{seed}{A whole number of the seed to use for random number generation
(or NULL). The final seed for each file is also based on the file name.}

\item{ask}{A flag specifying whether to ask before starting to process the files.}
}
\value{
An invisible named logical vector indicating for each file
whether it was successfully processed.
}
\description{
Starts (or restarts if previously stopped) processing the remaining files
specified by \code{\link{batch_config}()}.
}
\details{
\code{batch_run()} logs all file processing attempts together with the
the type (SUCCESS or FAILURE), the 
system time in UTC, the file name and any error messages.
The hidden log file can be read using \code{\link{batch_log_read}()}.

\code{\link{batch_files_remaining}()} provides a vector of the files that
are remaining to be processed.

When processing is complete the hidden configuration file
and hidden log file can be deleted using \code{\link{batch_cleanup}()}.

If a remaining file is removed or modified by a separate process,
\code{batch_run()} throws an error.
}
\seealso{
\code{\link{batch_process}()}, \code{\link{batch_config}()} and 
\code{\link{batch_cleanup}()}
}
