% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{batch_run}
\alias{batch_run}
\title{Runs Batch Processing}
\usage{
batch_run(path = ".", failed = FALSE, parallel = FALSE,
  logger = NULL)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{failed}{A logical scalar specifying whether to exclude (FALSE),
include (NA), or only include (TRUE) files that have thus far failed to process.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}

\item{logger}{A \code{\link[log4r]{logger}} object that is passed to 
the file processing function to allow the user to log their own messages.}
}
\value{
An invisible named logical vector indicating for each file
whether it was successfully processed.
}
\description{
Starts (or restarts if previously stopped) processing the files
specified in the configuration file.
}
\seealso{
\code{\link{batch_process}()}
}
