% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{batch_start}
\alias{batch_start}
\title{Start Batch Processing}
\usage{
batch_start(path, parallel = FALSE)
}
\arguments{
\item{path}{A string of the path to the directory with the files.}

\item{parallel}{A flag specifying whether to process the files in 
parallel (not yet used).}
}
\value{
An invisible character vector of the 
files processed by the current function call.
}
\description{
Starts (or restarts if stopped during processing) batch processing
files identified during batch set up.
}
