% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchr-package.R
\docType{package}
\name{batchr-package}
\alias{batchr}
\alias{batchr-package}
\title{batchr: Batch Process Files}
\description{
Batch processes files with a user-supplied R function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/batchr/}
  \item Report bugs at \url{https://github.com/poissonconsulting/batchr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)

Other contributors:
\itemize{
  \item Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca} [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
