% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchr-package.R
\docType{package}
\name{batchr-package}
\alias{batchr}
\alias{batchr-package}
\title{batchr: Batch Process Files}
\description{
Processes multiple files with a user-supplied
    function. The key design principle is that only files which were 
    last modified before the directory was configured are processed.
    A hidden file stores the configuration time and function etc while 
    successfully processed files are automatically touched to update 
    their modification date. As a result batch processing can be 
    stopped and restarted and any files created (or modified or deleted) 
    during processing are ignored.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://poissonconsulting.github.io/batchr/}
  \item Report bugs at \url{https://github.com/poissonconsulting/batchr/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Other contributors:
\itemize{
  \item Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca} [contributor]
  \item Poisson Consulting [copyright holder, funder]
}

}
\keyword{internal}
